refactor(cli): split src/cli.ts and CLI-only utils into src/cli/
Same shape as the plugin/ split (#218) but for the CLI tier: carves the
743-line src/cli.ts orchestrator and its 6 CLI-only utility files out of
src/utils/ and into a dedicated src/cli/ directory.

Files moved (via `git mv` so history is preserved):
  src/cli.ts                            -> src/cli/index.ts
  src/utils/annotation-encoding.ts      -> src/cli/annotation-encoding.ts
  src/utils/find-owning-project.ts      -> src/cli/find-owning-project.ts
  src/utils/get-staged-files.ts         -> src/cli/get-staged-files.ts
  src/utils/handle-error.ts             -> src/cli/handle-error.ts
  src/utils/parse-file-line-argument.ts -> src/cli/parse-file-line-argument.ts
  src/utils/select-projects.ts          -> src/cli/select-projects.ts

Picked these specifically because they are the only utils with cli.ts as
their sole non-test importer (per a classifier audit). Utils still used by
scan.ts, index.ts (public API), eslint-plugin.ts, or install-skill.ts
stay in src/utils/ since they aren't CLI-only.

Import path rewrites:
  - Inside the moved files: ./<other-cli-util>.js stays (same dir now),
    ./<shared-util>.js becomes ../utils/<util>.js
  - Inside cli/index.ts: ./utils/<x>.js -> ../utils/<x>.js for shared utils,
    -> ./<x>.js for the 6 moved CLI utils; ./scan.js, ./install-skill.js,
    ./constants.js, ./types.js -> ../<file>.js
  - 4 test files updated to import the moved utils from src/cli/ instead
    of src/utils/
  - vite.config.ts: cli entry "./src/cli.ts" -> "./src/cli/index.ts"

No behavior change. 734 tests pass, lint clean, typecheck clean, build
produces identical dist/cli.js output.
