+++ packages/react-doctor/src/cli/annotation-encoding.ts
