+++ packages/react-doctor/src/cli/find-owning-project.ts
@@ -1,5 +1,5 @@
 import path from "node:path";
-import { discoverReactSubprojects, listWorkspacePackages } from "./discover-project.js";
+import { discoverReactSubprojects, listWorkspacePackages } from "../utils/discover-project.js";
 
 export const findOwningProjectDirectory = (rootDirectory: string, filePath: string): string => {
   const absoluteFile = path.isAbsolute(filePath) ? filePath : path.resolve(rootDirectory, filePath);
