+++ packages/react-doctor/src/cli/get-staged-files.ts
